Clustering Analysis
In this project, I explored different clustering methods by applying K-means and hierarchical clustering on synthetic and real-world datasets. We evaluated the performance of the clustering algorithms using external validation metrics and visualized the data points in 2D or 3D plots, both for the original class and the class allocated by the clustering algorithm. This allowed us to compare and contrast the clustering methods and understand their strengths and weaknesses.

I used K-means and hierarchical clustering to group countries based on their world indicators, such as life expectancy, GDP, and infant mortality. We used internal validation metrics to assess the quality of the clustering and identified the best clustering solution. We then generated scatter plots to visualize the relationships between the world indicators and the clustering groups. This enabled us to gain insights into the similarities and differences between countries and understand how the clustering algorithms can be used for exploratory data analysis.

Overall, these projects provided me with a hands-on experience of working with clustering algorithms, evaluating their performance, and visualizing the results. I also learnt how clustering can be applied to real-world datasets to gain insights into complex data structures and identify patterns that can inform decision-making.